% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assesfitnessandqc.R
\name{assesfitnessandqc}
\alias{assesfitnessandqc}
\title{fitness and QC}
\usage{
assesfitnessandqc(link)
}
\arguments{
\item{link}{a link the DWC-A file (URL to IPT or a zip file)}
}
\description{
Function which returns codes refering to the fitness of use or an the quality
}
\examples{
\dontrun{
qc <- assesfitnessandqc("http://ipt.vliz.be/training/archive?r=biofun_2009")
}
}
