% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assesfitness.R
\name{assessfitness}
\alias{assessfitness}
\title{asses fitnes}
\usage{
assessfitness(Event = NULL, Occurrence = NULL, eMoF = NULL, output = list())
}
\arguments{
\item{Event}{optional a DwC event table}

\item{Occurrence}{optional a DwC occurrence table}

\item{eMoF}{optional a DwC eMoF table}

\item{output}{optional the output from importiptdata}
}
\description{
assess the fitness of an imported IPT file
}
\examples{
fitness <- assessfitness(output)
}
