% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BODCbiometrics.R
\docType{data}
\name{BODCbiometrics}
\alias{BODCbiometrics}
\title{BODC vocabulary terms related to Biometric or other Biotic data.}
\format{
An object of class \code{character} of length 12.
}
\usage{
BODCbiometrics
}
\description{
A character vector containing URIs of BODC vocabulary terms related to Biometric or other Biotic data. 
These will be added to th duplicates check if they exist in the dataset.
}
\keyword{datasets}
