% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BODCeffort.R
\docType{data}
\name{BODCeffort}
\alias{BODCeffort}
\title{BODC vocabulary terms related to sampling descriptors.}
\format{
An object of class \code{character} of length 30.
}
\usage{
BODCeffort
}
\description{
A character vector containing BODC vocabulary terms/collections IDs related to sampling descriptors.
}
\keyword{datasets}
