% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BODCinstrument.R
\docType{data}
\name{BODCinstrument}
\alias{BODCinstrument}
\title{BODC vocabulary term related to sampling instrument.}
\format{
An object of class \code{character} of length 1.
}
\usage{
BODCinstrument
}
\description{
A character string containing the BODC vocabulary term ID for "sampling instrument name".
}
\keyword{datasets}
