% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BODCnomofvalues.R
\docType{data}
\name{BODCnomofvalues}
\alias{BODCnomofvalues}
\title{BODC vocabulary term related to measurements or facts where no unitID or valueID apply.}
\format{
An object of class \code{character} of length 2.
}
\usage{
BODCnomofvalues
}
\description{
A character vector containing the BODC vocabulary terms URIs for measurements or facts where no UnitID or ValueID apply.
}
\keyword{datasets}
