% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BODCquantity.R
\docType{data}
\name{BODCquantity}
\alias{BODCquantity}
\title{BODC vocabulary term related to measurements of quantities where no valueID apply.}
\format{
An object of class \code{character} of length 20.
}
\usage{
BODCquantity
}
\description{
A character vector containing the BODC vocabulary terms URIs for measurements of quantities, used in the assesfitness function.
}
\keyword{datasets}
