% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BODCunits.R
\docType{data}
\name{BODCunits}
\alias{BODCunits}
\title{Copy of the BODC P06 collections.}
\format{
A data frame with updating number of rows and 4 columns:
\describe{
  \item{identifier}{ID of the vocab term}
  \item{preflabel}{preferred label of the vocab term}
  \item{altLabel}{alternative label of the vocab term}
  \item{uri}{uri of the vocab term}
}
}
\source{
\url{http://vocab.nerc.ac.uk/collection/}
}
\usage{
BODCunits
}
\description{
A dataframe containing the BODC P06 collections. These vocabulary terms can be used as measurementUnitIDs.
}
\keyword{datasets}
