% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BODCvalues.R
\docType{data}
\name{BODCvalues}
\alias{BODCvalues}
\title{Copy of the BODC L22, L05, F02, C17, S11, S10 and M20 collections and the EUNIS collection.}
\format{
A data frame with updating number of rows and 3 columns:
\describe{
  \item{definition}{definition of the vocab term}
  \item{preflabel}{preferred label of the vocab term}
  \item{deprecated}{states if the term is deprecated}
  \item{uri}{uri of the vocab term}
}
}
\source{
\url{http://vocab.nerc.ac.uk/collection/}
}
\usage{
BODCvalues
}
\description{
A dataframe containing the BODC L22, L05, F02, C17, S11, S10 and M20 collections and the EUNIS collection. 
These vocabulary terms can be used as measurementValueIDs.
}
\keyword{datasets}
