% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataset2.R
\name{checkdataset2}
\alias{checkdataset2}
\title{Check the quality of the dataset}
\usage{
checkdataset2(Event = NULL, Occurrence = NULL, eMoF = NULL,
  IPTreport = list(), tree = FALSE)
}
\arguments{
\item{Event}{optional parameter, the name of the event core file}

\item{Occurrence}{mandatory parameter, the name of the occurrence file}

\item{eMoF}{optional parameter, the name of the mof or emof file}

\item{IPTreport}{optional parameter, in case you want to append the result to an existing listfile}

\item{tree}{optional parameter, takes value yes if you want the QC report to include the OBIS tree hierachy}
}
\description{
This function does all QC checks of the package
}
\examples{
IPTreport <-checkdataset(Event = event, Occurrence = occurrence, eMoF = emof, IPTreport = IPTreport, tree = FALSE)
}
