% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIPTdataset.R
\name{checkIPTdataset}
\alias{checkIPTdataset}
\title{Check the quality of the dataset}
\usage{
checkIPTdataset(link, tree = FALSE)
}
\arguments{
\item{link}{mandatory parameter, link to an IPT resource or an DwC zip file}

\item{tree}{optional parameter, takes value yes if you want the QC report to include the OBIS tree hierachy}
}
\description{
This function takes a DWC file as input and carries out the checkdataset funtion. The output is a list.
}
\examples{
IPTreport <- checkIPTdataset(link = "http://ipt.vliz.be/upload/resource?r=biofun2009", tree = FALSE)
IPTreport <- checkIPTdataset(link = "dwca-biofun2009-v1.1.zip", tree = "yes")
}
