% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfunctions.R
\name{cleandataframe}
\alias{cleandataframe}
\title{remove empty collumns and make all collumns into characters}
\usage{
cleandataframe(x, vector = TRUE)
}
\arguments{
\item{x}{a dataframe}

\item{vector}{A logical scalar. If TRUE x will be transformed to character dataframe}
}
\description{
remove empty collumns and make all collumns into characters
}
