% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_checks.R
\name{date_in_future}
\alias{date_in_future}
\title{Identifies dates that are in the future.}
\usage{
date_in_future(x)
}
\arguments{
\item{x}{mandatory. a dataframe that contains a field called eventDate}
}
\description{
This function highlights dates that are in the future and flag them as errors
}
\examples{
\dontrun{
output <- suspicious_old_date(x = Occurrence)
output <- suspicious_old_date(x = Event)
}
}
