% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportIPT.R
\name{exportIPT}
\alias{exportIPT}
\title{Create QC report in markdown}
\usage{
exportIPT(IPTreport)
}
\arguments{
\item{IPTreport}{mandatory parameter, the output from checkdataset or checkIPTdataset}
}
\description{
This function takes as input the output from checkdataset and checkIPTdataset and generates a QC report in markdown.
}
\examples{
exportIPT(IPTreport)
}
