% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncols.R
\name{fncols}
\alias{fncols}
\title{fncols}
\usage{
fncols(data, cname)
}
\arguments{
\item{data}{mandatory parameter, name dataframe}

\item{cname}{mandatory parameter, collumn name}
}
\description{
checks if a collumn exists in the dataframe and adds one if not.
}
