% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getskosxmldata.R
\name{getskossxmldata}
\alias{getskossxmldata}
\title{get skos xml data}
\usage{
getskossxmldata(vocids)
}
\arguments{
\item{vocids}{mandatory parameter, the XML link(s)  you want to read the read}
}
\description{
This function allows to get the information form different concepts stored in skoss xml. It takes a single concept, multiple concepts or a shema. The output is a list of 2 data frames; one with the lables and definitions. The other with the relations with other concepts.
}
\examples{
getskossxmldata("http://vocab.nerc.ac.uk/collection/P06/current/UMCU/")
getskossxmldata(c("http://vocab.nerc.ac.uk/collection/P06/current/UMCU/","http://vocab.nerc.ac.uk/collection/P06/current/ULAA/")
getskossxmldata("http://vocab.nerc.ac.uk/collection/S10/current/")
getskossxmldata("http://dd.eionet.europa.eu/vocabulary/biodiversity/eunishabitats/rdf")
}
