% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importiptdata.R
\name{importiptdata}
\alias{importiptdata}
\title{import IPT datasets}
\usage{
importiptdata(file)
}
\arguments{
\item{file}{mandatory a link the DWC-A file (URL to IPT or a zip file)}
}
\description{
imports and cleans IPT datasets
}
\examples{
output <- importiptdata("http://ipt.vliz.be/training/archive?r=biofun_2009")
}
