% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopcheckIPTdataset.R
\name{loopcheckIPTdataset}
\alias{loopcheckIPTdataset}
\title{Check the quality of multiple dataset}
\usage{
loopcheckIPTdataset(x, tree = FALSE)
}
\arguments{
\item{x}{mandatory parameter, a vector with one or more IPT links or DwC file locations}

\item{tree}{optional parameter, takes value yes if you want the QC report to include the OBIS tree hierachy}
}
\description{
This function takes multiple ipt or dwc files as input and generates html QC reports
}
\examples{
\dontrun{
loopcheckIPTdataset(c("http://ipt.vliz.be/upload/resource?r=biofun2009","http://ipt.vliz.be/upload/resource?r=ukranianromanian_benthos_blacksea_iaea_1998"))
}
}
