% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopfitnesandqc.R
\name{loopfitnessandqc}
\alias{loopfitnessandqc}
\title{fitness and QC loop}
\usage{
loopfitnessandqc(links)
}
\arguments{
\item{mandatory}{a link the DWC-A file (URL to IPT or a zip file)}
}
\description{
loops fitness and qc function
}
\examples{
qc <- assesfitnessandqc("http://ipt.vliz.be/training/archive?r=biofun_2009")
}
