% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfunctions.R
\name{midstring}
\alias{midstring}
\title{find and select charcters in a string}
\usage{
midstring(x, first, last, n = 0, m = 0)
}
\arguments{
\item{x}{mandatory parameter, the input string}

\item{first}{mandatory parameter, the character before the text you need}

\item{last}{mandatory parameter, the character after the text you need}

\item{n}{optional paramter, the offset to first}

\item{m}{optional paramter, the offset to last}
}
\description{
Function and select charcters in a string
}
