% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namestovector.R
\name{NamesToVector}
\alias{NamesToVector}
\title{convert IPT names to vector}
\usage{
NamesToVector(x)
}
\arguments{
\item{x}{mandatory parameter, input}
}
\description{
Function used by loopcheckIPTdataset to convert the imput to a vector.
}
