% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p01todwc.R
\docType{data}
\name{p01todwc}
\alias{p01todwc}
\title{BODC P01 vocab terms equivalent to some DwC terms.}
\format{
A data frame with updating number of rows and 2 columns:
\describe{
  \item{P01s}{BODC P01 collection ID of a certain term}
  \item{dwc}{DwC field name equivalent to the P01 term}
}
}
\usage{
p01todwc
}
\description{
A dataframe containing a number of BODC P01 vocab terms equivalent to some DwC terms. .
}
\keyword{datasets}
