% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reversetaxmatch.R
\name{reversetaxmatch}
\alias{reversetaxmatch}
\title{reversed taxonmatch}
\usage{
reversetaxmatch(aphiaIDs)
}
\arguments{
\item{aphiaIDs}{mandatory parameter, a vector with aphiaIDs from WoRMS}
}
\description{
adds the classification from worms to a vector with aphiaIDs
}
\examples{
\dontrun{
reversetaxmatch(aphiaIDs)
}
}
