% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skossconcepts.R
\docType{data}
\name{skossconcepts}
\alias{skossconcepts}
\title{SKOS concept types.}
\format{
An object of class \code{character} of length 4.
}
\usage{
skossconcepts
}
\description{
A character vector containing SKOS concept types.
}
\keyword{datasets}
