% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_checks.R
\name{suspicious_old_date}
\alias{suspicious_old_date}
\title{Identifies dates that are older than a threshold and flag them as suspicious.}
\usage{
suspicious_old_date(x)
}
\arguments{
\item{x}{mandatory. a dataframe that contains a field called eventDate}
}
\description{
This function highlights dates that are very old
}
\examples{
\dontrun{
output <- suspicious_old_date(x = Occurrence)
output <- suspicious_old_date(x = Event)
}
}
