% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfunctions.R
\name{txt_to_cols}
\alias{txt_to_cols}
\title{txt_to_cols separates a one-column dataframe into a multiple column dataframe based on a separator}
\usage{
txt_to_cols(x, sep = ",")
}
\arguments{
\item{x}{dataframe to transform}

\item{sep}{separator to be used to separate the dataframe into multiple columns}
}
\description{
txt_to_cols separates a one-column dataframe into a multiple column dataframe based on a separator
}
