% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallfunctions.R
\name{valid_url}
\alias{valid_url}
\title{Checks if an url exist}
\usage{
valid_url(url_in, t = 2)
}
\arguments{
\item{url_in}{mandatory parameter, the input string}

\item{t}{optional parameter, the timeout of the open.connection}
}
\description{
Checks if an url exist better than the url.exist function
}
