% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODVtoDwC.R
\name{ODVtoDwC}
\alias{ODVtoDwC}
\title{ODV to DwC}
\usage{
ODVtoDwC(file)
}
\arguments{
\item{file}{mandatory parameter, the location of the .txt file you want to convert}
}
\description{
converts ODV files exported from ODV to DwC (occurrence and eMoF tables)
}
\examples{
ODVtoDwC("data_from_collection_MinimumObservationDepth.txt")
}
